+++ vim-engine/src/main/kotlin/com/maddyhome/idea/vim/api/stubs/VimProcessGroupStub.kt
@@ -9,26 +9,14 @@
 package com.maddyhome.idea.vim.api.stubs
 
 import com.maddyhome.idea.vim.KeyProcessResult
-import com.maddyhome.idea.vim.api.ExecutionContext
 import com.maddyhome.idea.vim.api.VimEditor
 import com.maddyhome.idea.vim.api.VimProcessGroupBase
-import com.maddyhome.idea.vim.command.Command
 import com.maddyhome.idea.vim.diagnostic.vimLogger
 import javax.swing.KeyStroke
 
 class VimProcessGroupStub : VimProcessGroupBase() {
   init {
     vimLogger<ExecutionContextManagerStub>().warn("VimProcessGroupStub is used. Please replace it with your own implementation of VimProcessGroup.")
-  }
-
-  override fun startExEntry(
-    editor: VimEditor,
-    context: ExecutionContext,
-    command: Command,
-    label: String,
-    initialCommandText: String
-  ) {
-    TODO("Not yet implemented")
   }
 
   override fun processExKey(editor: VimEditor, stroke: KeyStroke, processResultBuilder: KeyProcessResult.KeyProcessResultBuilder): Boolean {
